Fix reference to enable_js_type_check
A TODO in ninjalog_uploader.py referred to the closure_compile gn arg
which was recently renamed (as of crrev.com/c/2248564) to
enable_js_type_check. This updates it (and adds a missing space and a
bug number). This is purely a comment fix.

Change-Id: I117fa7910663491fd7c6e017fe98e2a977a75f13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/2336176
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>