[man] Fix link to the git repository
"here" is not a nice name for the link, and the footnote generated for
the man page. Replace it by a properly named link.

Note that some of the changes in the generated files are unrelated.
Those files have not been updated since May 2019, and it seems like some
previous changes were not reflected yet.

R=ehmaldonado@chromium.org

Change-Id: I31c784e7eb649bcaa74ba223027fb2c142f52838
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/2346371
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Clemens Backes <clemensb@chromium.org>