Remove tmp gclient files on fetch
Regression was introduced as part of gclient_scm change to support cog.
The directory should be cleaned up regardless if is empty or not, and if
statement is used only to warn user about non-empty directory, which si
not expected.

R=jojwang@google.com

Fixed: 1339077
Change-Id: Ie3bc6bcac168bad9cf2cb6adc6ae4024fe9ef167
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/3724051
Commit-Queue: Joanna Wang <jojwang@chromium.org>
Auto-Submit: Josip Sokcevic <sokcevic@google.com>
Reviewed-by: Joanna Wang <jojwang@chromium.org>