[apple] Set limit of 200 process when running on macOS Ventura 13.5
The limit set in /Library/LaunchDaemons/limit.maxfiles.plist appears
to no longer be respected in macOS Ventura 13.5 and the OS forces a
limit of 256 file descriptors, causing build with autoninja to fail
on macOS running this most recent version of the OS.

Force the max number of process to 200 until a new working way to
increase the file descriptor limit is found. This will allow devs
working on macOS to be able to build (albeit slower than before).

Fixes the following build failure when goma is enabled and running
on macOS Ventura 13.5:

    $ autoninja -C out/Debug-iphonesimulator chrome
    ninja: Entering directory `out/Debug-iphonesimulator'
    [0/3574] CXX ....oninja: fatal: pipe: Too many open files

Bug: 1467777
Change-Id: Ia7eaab552f7e6d26a2f48d72bb8235a70d6d442f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/4720227
Reviewed-by: Dirk Pranke <dpranke@google.com>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>