cros: expand to cover more roots
The background is we maintain a fork of a fork of cros and as we're
moving to submodules this is what our repo looks like e.g. chromite is
at the top, and it's using submodules.

It doesn't seem like this would negatively affect anyone else, and would
make our system a lot simpler.

Change-Id: Ib1c9fd95a36e55fc13bede9bb539daedd5efe0ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/5132338
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>