Add config to the telemetry lib
This provides the consistent config between runs. This only supports
linux at the moment, a windows support CL will follow up. Note the
current count down is stored to ensure we provide the banner a minimum
of 10 times and a uuid is stored which is collected in lieu of the
actual username

Bug: 326277821
Change-Id: I3817436ea213c755f490914023508a10cf3cef1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/5855459
Reviewed-by: Terrence Reilly <treilly@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>