[gclient_scm] Delete auto configure in gclient
This causes more trouble than it's worth (it's breaking dawn builds).

That does mean for a fresh depot_tools install, users will need to
explicitly run git cl creds-check once, which means tracking down all
of the docs that need to say that.

In the meantime, it's probably better to delete this now.

Bug: 408427309
Change-Id: I5595756441500c43c543b47d97ff6b1bd9ddab30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6434734
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>