cipd: ignore .cipd_client_root
It was introduced to override the cipd bin path so that the downloaded
files wouldn't be listed in the CiderG workspace file tree when
the bundled depot_tools is used.

However,
- chromium no longer uses the bundled depot_tools
- the cipd_client_root, pointing to a folder in /tmp, causes issues
when a temp cleanup cron removes files based on the last access ts.

We no longer need .cipd_client_root.

Bug: 415080055
Change-Id: Ic966699cce3173a8ce4033752edec5e53aded870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6614519
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>