Fix setbotcommit comment on ambiguous revision
Add revision number option to gerrit_client.py setbotcommit command. Add
an explicit revision number to a setbotcommit call in the update_files
method.

After creating a new patch-set in the release recipe and verifying it
was created we are adding a comment which could be applied to the
previous patch. Setting `Bot-Commit+1` relied on the default 'current'
revision which may be stale in some Gerrit replicas. This CL applies the
fix suggested in 412739202

Bug: 414681998
Change-Id: I1d380e221864bcf1183c77fa58be4ede9a787ef1
Recipe-Nontrivial-Roll: build_internal
Recipe-Nontrivial-Roll: chrome_release
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6718447
Reviewed-by: Alex Kravchuk <alexanderkr@google.com>
Reviewed-by: Scott Lee <ddoman@chromium.org>