Correctly accumulate gs upload URLs
https://crrev.com/c/6796221 added a gsutil_urls output property that
provides the URLs of uploads to Google storage that were performed by
the build. It attempts to accumulate the URLs by updating a dictionary
on the result object's properties, but the properties on a step's
results object is empty when created, so it was just overriding the
property each time with a single element dict. This change keeps the
dict on the API object and updates it when an upload is performed so
that the final value includes all of the URLs produced by the build.

Change-Id: I6681ca7c137969cbef58e09ce24d0605155d2c3e
Recipe-Nontrivial-Roll: build
Recipe-Nontrivial-Roll: build_internal
Recipe-Nontrivial-Roll: chrome_release
Recipe-Nontrivial-Roll: chromiumos
Recipe-Nontrivial-Roll: infra
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6880970
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>